Created script to mint in batch
