+++ constants/index.ts
@@ -1,0 +1,1 @@
+export const SLOPE = "0.00005";
