+++ package.json
@@ -47,6 +47,7 @@
     "typescript": "^4.3.4"
   },
   "dependencies": {
+    "csv-writer": "^1.6.0",
     "hardhat-abi-exporter": "^2.8.0"
   }
 }
