+++ test/mint_total_supply.spec.ts
@@ -1,13 +1,15 @@
 import { ethers } from "hardhat";
 
+import { SLOPE } from "../constants";
+
 const factoryAddress = "0x41C659319885598d77CF5bd8E792A5162bC72A04";
 const mintParams = {
-  creator: "0x55D0ba6789A7866b0e0dc100b1B6F1Da0361bCD7",
+  creator: "0x59376b45300fFa73FebF43C0932307FaC426E7EC",
   usdcCollateral: "0x63aF7615e795F2cFb8A2f93aFAd7CD1B4d35bA5c",
   usdtCollateral: "0xb34Ca2cDE88dE520E4Be8b1ccEc374D3052ae021",
-  slope: ethers.utils.parseEther("0.0005").toString(),
+  slope: ethers.utils.parseEther(SLOPE).toString(),
   maxSupply: ethers.utils.parseEther("10000000").toString(),
-  name: "test_artist_5",
+  name: "test_artist_15",
   symbol: "TESTARTIST",
 };
 
@@ -15,7 +17,7 @@ describe("Mint total supply", async () => {
   let token = "";
   let signers: any;
   before(async () => {
-    const whiteAddress = "0x55D0ba6789A7866b0e0dc100b1B6F1Da0361bCD7";
+    const whiteAddress = "0x59376b45300fFa73FebF43C0932307FaC426E7EC";
     signers = await ethers.getSigners();
 
     const [owner] = await ethers.getSigners();
